Move node id assigning into `resolve`
